feat: hide alert when navigating to protect wallet (#7534)
**Description**

This PR aims to solve a bug on the backup alert, when navigating to the
protect your wallet screen the alert kept being visible.

This PR contains BackupAlert converted to typescript and functional
component.

## **Manual testing steps**

_1. Step1: Create a new wallet
_2. Step2: Do not protect your wallet pressing (remember later)
_3. Step3: Press protect your wallet on the backup alert
_4. Step4: The protect your wallet should not be there

## **Screenshots/Recordings**
https://recordit.co/skV2sxdzil

new Bell Icon:

![image](https://github.com/MetaMask/metamask-mobile/assets/46944231/71b95e66-8d71-4951-a02e-4beca3e07af5)

![image](https://github.com/MetaMask/metamask-mobile/assets/46944231/7a9afc84-4138-4d91-9f8e-4f564b757125)

### **Before**

_[screenshot]_

### **After**

_[screenshot]_

## **Related issues**

_Fixes #4400 

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've clearly explained:
  - [ ] What problem this PR is solving.
  - [ ] How this problem was solved.
  - [ ] How reviewers can test my changes.
- [ ] I’ve indicated what issue this PR is linked to: Fixes #???
- [ ] I’ve included tests if applicable.
- [ ] I’ve documented any added code.
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------
