+++ app/components/UI/BackupAlert/BackupAlert.constants.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const PROTECT_WALLET_BUTTON = 'protect-your-wallet-button';
