+++ app/components/UI/BackupAlert/BackupAlert.types.ts
@@ -1,0 +1,4 @@
+export interface BackupAlertI {
+  navigation: any;
+  onDismiss: () => void;
+}
