+++ app/components/UI/BackupAlert/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BackupAlert';
