+++ package.json
@@ -419,6 +419,7 @@
     "@types/react-native": "^0.64.10",
     "@types/react-native-actionsheet": "^2.4.3",
     "@types/react-native-background-timer": "^2.0.0",
+    "@types/react-native-elevated-view": "^0.0.4",
     "@types/react-native-material-textfield": "^0.16.5",
     "@types/react-native-scrollable-tab-view": "^0.10.3",
     "@types/react-native-svg-charts": "^5.0.12",
